requirements.txt
facebook_business==21.0.0
python-dotenv==1.0.0
# google-generativeai==0.3.0  # Uncomment when Gemini integration is implemented
